Backend/requirements.txt
fastapi
uvicorn
python-multipart
moviepy
openai-whisper
transformers
torch
passlib[argon2]
python-jose
python-dotenv
imageio[ffmpeg]
pymongo
email-validator
requests
pymongo[bson]
jinja2